refactor admin list routes
