+++ webapi/src/test/resources/logback-test.xml
@@ -134,7 +134,7 @@
     <logger name="org.knora.webapi.responders.admin.ListsAdminResponderSpec" level="INFO"/>
     <logger name="org.knora.webapi.responders.admin.StoresResponderADM" level="INFO"/>
     <logger name="org.knora.webapi.responders.admin.UsersResponderADMSpec" level="INFO"/>
-    <logger name="org.knora.webapi.e2e.admin.lists.OldListsRouteADMFeatureE2ESpec" level="INFO"/>
+    <logger name="org.knora.webapi.e2e.admin.lists.ListsRouteADME2ESpec" level="INFO"/>
     <logger name="org.knora.webapi.e2e.admin.ProjectsADME2ESpec" level="INFO"/>
     <logger name="org.knora.webapi.e2e.admin.StoreADME2ESpec" level="INFO"/>
     <logger name="org.knora.webapi.e2e.admin.UsersADME2ESpec" level="INFO"/>
